Switch from a process plugin to a WASM plugin
sqlc supports process and WASM plugins. A process plugin causes sqlc to spawn a child process to execute plugin code, while a WASM plugin is executed within the sqlc process.

Our previous approach to generating Kotlin code had us writing JVM code for us in a process plugin. We had sqlc spawn a fresh JVM to execute our code. This was slow and tedious to set up, because sqlc checks if the process command exists as a file. So "java -cp ..." wouldn't work because that's not a file. We looked into publishing a Gradle distribution ZIP, but that too was tedious and still leaves us with a slow code generation step.

Instead, we wrote a tiny Rust program that dumps the sqlc data into a JSON file. We package this program as a WASM plugin, so it's really fast. We then consume the JSON file and generate the Kotlin code from it, in our Gradle plugin. The Gradle JVM is already running, so this step is also fast. Plus, the code can now be debugged, and we get good stack traces when there is a problem. JSON was chosen instead of protobuf to aid in manual understanding, bug reporting, etc.

The change required a heavy rewrite of the Gradle plugin code. We used this opportunity to combine some of the old and new code, though mostly the old was kept. The Gradle plugin is now also functional, at least on Mac.